Move code only used in GridSaveAs (unshipped utility script) to scripts (archive) folder.
